Move raw implementations into `raw` module
